Cleanup and correction of extreme gear data.

+++ Sewer56.SonicRiders/Structures/Gameplay/ExtremeGearLevelStats.cs
@@ -25,7 +25,7 @@ namespace Sewer56.SonicRiders.Structures.Gameplay
 
         public int BoostCost;
         public int TornadoCost;
-        public int SpeedGainedFromDriftDash;
-        public int BoostSpeed;
+        public float SpeedGainedFromDriftDash;
+        public float BoostSpeed;
     }
 }
